+++ src/lib/Crud/definition.ts
@@ -1,6 +1,7 @@
 import type { CrudOperation } from '$lib/Crud/Operations';
 import type { StateProvider } from '$lib/State/Provider';
 import type { StateProcessor } from '$lib/State/Processor';
+import {type DashboardDefinition} from "$lib/Dashboard/definition";
 
 export type CrudDefinitionOptionsArgument<T> = {
 	name: string,
@@ -22,6 +23,7 @@ export type CrudDefinitionOptions<T> = Required<CrudDefinitionOptionsArgument<T>
 export class CrudDefinition<T> {
 	public readonly name: string;
 	public readonly options: CrudDefinitionOptions<T>;
+	private _dashboard: DashboardDefinition|null = null;
 
 	constructor(options: CrudDefinitionOptionsArgument<T>) {
 		const name = options.name;
@@ -50,10 +52,28 @@ export class CrudDefinition<T> {
 					.join(', ')}.`
 			);
 		}
+		options.operations.forEach((operation: CrudOperation) => operation.crud = this);
 
 		options.defaultOperationName = defaultOperation.name;
 		options.identifierFieldName ??= 'id';
 
 		this.options = options as CrudDefinitionOptions<T>;
+	}
+
+	get dashboard(): DashboardDefinition {
+		if (!this._dashboard) {
+			throw new Error('Dashboard is not set in Crud definition: did you try to bypass Crud setup?');
+		}
+		return this._dashboard;
+	}
+
+	set dashboard(dashboard: DashboardDefinition) {
+		if (this._dashboard) {
+			throw new Error('Cannot set dashboard twice in a Crud definition.');
+		}
+		this.options.operations.forEach((operation) => {
+			operation.dashboard = dashboard;
+		});
+		this._dashboard = dashboard;
 	}
 }